[Ganesh] - added the change colour feature
What changes have you made?
  when the user selects to play a change colour card.. the game will
check whether it have a running colour card(only for +4) if the player
have then the he wont allow to play the card... if he dont then the game
will show a change colour menu.. which will help the user to select the
new colour

Why do we need this change?
  to allow the user to make the move if he dont have the running colour