[Ganesh] - fixed the bug from draw card
What changes have you made?
  when the player makes a request to draw the card... after drawing he
can only play that card or he can pass the turn to the next player...but
previously the player was able to draw multiple cards in one chance.. so
fixed that bug...now a player can draw only one card at a time...and
then he can play that card or can pass the turn to the next player..
*

Why do we need this change?
  chnages needed for the fair play of the game..
*